Gradient Ascent, Hill-Climbing - Complex Landscape
Implementation of 2 popular fitness function / optimization algorithms. 

Gradient Ascent algorithm starts from a random position in a given search-space, takes a step in the steepest direction uphill and repeats until a global optimum is reached. Since the gradient points in the steepest direction uphill the algorithm, we will evaluate the gradient at each iteration. We apply a variable learning rate and stopping condition. 

Hill-Climbing algorithm is more stochastic than Gradient Ascent. We use a stopping condition, a mutation process and rate of mutation. We then continue climbing the search-space until a global optimum is reached. In contrast to Gradient Ascent, Hill-Climbing algorithm will less likely settle on local optimum as a result of mutation processes and nearest-neighbor search logic.